data loading with changes for speed
JJD put this together and it includes changes that drastically improved run-time for the 1 area model. The decision was made to stick with 7 fleets and some of the changes JJD made are hard coded, and may have broken the functionality of the fleetcombo option available in the other data_loading script. So this script is somewhat redundant and could be fixed and get us back to only a single data_loading script, instead of this mish-mash of similar yet different scripts. In terms of changes that effected run-time and model performance, pay close attention to phases and penalties. Nearly all penalties should be off, except maybe an early F pen, but even that turns off for final estimation.